Add ability to specify additional conversion options
This is possible by adding an additional Map<String,String> to startConversions.
The Options are specified in form format:

```
options.put("converteroptions[templating][username]", "Hans");
```

Optimally one could specify an Object that gets converted to JSON, but I wasn't able getting the API to accept the JSON in combination with the multipart form upload.